BluetoothInCallService: Fix '+' display in LE Audio calls
In `toLeCall()`, `Uri.toString()` was causing the dialing prefix
'+' to be encoded as '%2B', resulting in incorrect caller ID
displaying on LE Audio devices.

For example, if the URI was `tel:+12345`, in the notification
sent to the remote it would display as `tel:%2B12345`.

This CL switches `toString() to `getScheme()` combined with
`getSchemeSpecificPart()` to pass the unencoded URI and
changes toLeCall() visibility for more robust testing.

Therefore after this change, `tel:+12345` will display
exactly as `tel:+12345`.

Test: atest BluetoothInCallServiceTest
Test: manual PTS test: GTBS/SR/CP/BV-01-C
Bug: 447308191
Flag: EXEMPT BUGFIX
Change-Id: Ie30da3995e3906cd9f46fbe60b4ad319c34f1fb4