Change device type verification logic in BTM
We currently read AD type flags from the advertising reports to
determine which device type the remote is. This process in BTM and GD is
different, so based on information in the Bluetooth specification,
unifying the process to that of GD. We simply read the
'BREDR_NOT_SUPPORTED` flag, and if it is not set to 1, we don't assume
anything and mark the device type as dual mode.

Bug: 471319424
Bug: 463502688
Test: m -j
Flag: com.android.bluetooth.flags.unify_device_type_verification_logic
Change-Id: I8801be0b30c74916913eb3fd8cc829f6cf2ba8c7