Make AvrcpControllerVolumeHandler objects one time use
By making AvrcpControllerVolumeHandler objects one time use (created
and destroyed for the Connected state), we avoid start-stop lifecycle
issues. However, this would mean we cannot log the desired volume
strategy unless the remote device is in the Connected state.

The desired strategy selection logic can be moved into a static method
because it is a function of only the sink device. This can then be
logged in AvrcpControllerService's dump(), even when remote devices are
not Connected.

Bug: 453682934
Flag: EXEMPT PURE_REFACTOR
Test: atest AvrcpControllerStateMachineTest
Test: atest AvrcpControllerVolumeHandlerTest
Test: atest AvrcpControllerServiceTest
Change-Id: I6b5a0ba5382044b6ea54429da0094dd9c6f01bd1